// src/clientscript/definitions.d.ts
export declare const namedClientScriptOps: {
    pushint: number;
    pushlong: number;
    pushstring: number;
    pushlocalint: number;
    poplocalint: number;
    pushlocalstring: number;
    poplocalstring: number;
    pushlocallong: number;
    poplocallong: number;
    popdiscardint: number;
    popdiscardlong: number;
    popdiscardstring: number;
    joinstring: number;
    gosub: number;
    pushvar: number;
    popvar: number;
    pushvarbit: number;
    popvarbit: number;
    jump: number;
    branch_not: number;
    branch_eq: number;
    branch_lt: number;
    branch_gt: number;
    branch_lteq: number;
    branch_gteq: number;
    branch_eq_long: number;
    branch_not_long: number;
    branch_lt_long: number;
    branch_gt_long: number;
    branch_lteq_long: number;
    branch_gteq_long: number;
    switch: number;
    return: number;
    pushconst: number;
    tribyte1: number;
    tribyte2: number;
    shorting_or: number;
    shorting_and: number;
    plus: number;
    minus: number;
    intdiv: number;
    intmul: number;
    strcmp: number;
    strconcat: number;
    inttostring: number;
    enum_getvalue: number;
    struct_getparam: number;
    item_getparam: number;
    quest_getparam: number;
    npc_getparam: number;
    cc_getparam: number;
    mec_getparam: number;
    dbrow_getfield: number;
    dbrow_findnext: number;
    cc_setparam: number;
    db_find_with_count: number;
    lc_getparam: number;
    enum_getstring: number;
    enum_getreverseindex: number;
    enum_getreversecount: number;
    enum_hasoutput: number;
    define_array: number;
    pop_array: number;
    push_array: number;
    printmessage: number;
};
// ...
export type PrimitiveType = "int" | "long" | "string";
// ...
export declare const branchInstructionsOrJump: number[];
// ...
export type SwitchJumpTable = {
    value: number;
    jump: number;
}[];
export type ClientScriptOp = {
    opcode: number;
    imm: number;
    imm_obj: string | number | [number, number] | {
        type: "switchvalues";
        value: SwitchJumpTable;
    } | {
        type: "jumplabel";
        value: ClientScriptOp;
    } | null;
    opname?: string;
};
// ...
export type StackConst = ClientScriptOp["imm_obj"];
export type StackType = PrimitiveType | "vararg";
export type StackTypeExt = StackType | StackDiff;
export declare class StackList {
    values: StackTypeExt[];
    constructor(values?: StackTypeExt[]);
    pushone(type: StackType): void;
    int(): void;
    long(): void;
    string(): void;
    isEmpty(): boolean;
    total(endoffset?: number): number;
    tryShift(n: number): boolean;
    hasSimple(other: StackList): boolean;
    pop(list: StackList, limit?: number): void;
    tryPopReverse(list: StackList, limit?: number): number;
    tryPopUnordered(otherval: StackDiff): boolean;
    tryPopSingle(otherval: StackType): boolean;
    tryPop(list: StackList, limit?: number): number;
    push(list: StackList): void;
    clone(): StackList;
    toString(): string;
    toFunctionBindString(): string;
    toTypeScriptVarlist(withnames: boolean, withtypes: boolean, exacttype?: ExactStack | null): string;
    toTypeScriptReturnType(exacttype?: ExactStack | null): string;
    toJson(): (StackDiff | StackType)[];
    static fromJson(v: ReturnType<StackList["toJson"]>): StackList;
    getStackdiff(): StackDiff;
    toStackDiff(): StackDiff;
    toLooseSubtypes(): number[];
}
export declare class ExactStack {
    int: number[];
    long: number[];
    string: number[];
    all(): number[];
    static fromList(types: number[]): ExactStack;
    static fromJson(json: ReturnType<ExactStack['toJson']>): ExactStack;
    toJson(): this;
}
export declare class StackInOut {
    in: StackList;
    out: StackList;
    exactin: ExactStack | null;
    exactout: ExactStack | null;
    constout: StackConst;
    initializedin: boolean;
    initializedout: boolean;
    initializedthrough: boolean;
    constructor(inlist?: StackList, outlist?: StackList);
    static fromExact(inlist: number[], outlist: number[]): StackInOut;
    getBottomOverlap(): number;
    totalChange(): number;
    getCode(): string;
    toString(): string;
    toJson(): {
        in: (StackDiff | StackType)[];
        out: (StackDiff | StackType)[];
        initializedthrough: boolean;
        exactin: ExactStack | undefined;
        exactout: ExactStack | undefined;
    };
    static fromJson(json: ReturnType<StackInOut["toJson"]>): StackInOut;
}
export declare class StackDiff {
    int: number;
    long: number;
    string: number;
    vararg: number;
    static fromJson(json: ReturnType<StackDiff["toJson"]> | undefined | null): StackDiff | null;
    toJson(): this;
    constructor(int?: number, long?: number, string?: number, vararg?: number);
    sub(other: StackDiff): this;
    add(other: StackDiff): this;
    min(other: StackDiff): void;
    max(other: StackDiff): void;
    mult(n: number): this;
    intdiv(n: number): this;
    lteq(other: StackDiff): boolean;
    equals(other: StackDiff): boolean;
    isEmpty(): boolean;
    isNonNegative(): boolean;
    toString(): string;
    total(): number;
    clone(): StackDiff;
    isMonoType(): PrimitiveType | "multi";
    getSingle(stack: StackType): number;
    setSingle(stack: StackType, value: number): void;
    getArglist(): StackList;
    toFunctionBindString(): string;
    toTypeScriptVarlist(nameoffset: StackDiff, withnames: boolean, withtypes: boolean, exacttype?: ExactStack | null): string;
}